scripts/requirements.txt
# Python dependencies for image fetching script
# Install with: pip install -r scripts/requirements.txt

pillow>=10.0.0
requests>=2.31.0

